Authentication on ProductController

+++ app/Http/Controllers/ProductController.php
@@ -6,35 +6,39 @@ use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
 use App\Http\Resources\ProductCollection;
 use App\Http\Resources\ProductResource;
+use App\Models\Admin;
+use App\Models\Customer;
 use App\Models\Product;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
+    public function __construct() {
+        $this->middleware('auth:api');
+    }
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request)
     {
+    
         $products = Product::with('category')->get();
-
         return new ProductCollection($products);
 
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
 
     /**
      * Store a newly created resource in storage.
      */
     public function store(StoreProductRequest $request)
     {
+        $admin =  Admin::where('user_id', Auth::user()->id)->first();
+        if(empty($admin)){
+         return response()->json(['message' => 'NO AUTHENTICATED'],201);
+        }
         $product = Product::create($request->validated());
         return new ProductResource($product);
     }
@@ -54,6 +58,10 @@ class ProductController extends Controller
      */
     public function update(UpdateProductRequest $request, Product $product)
     {
+        $admin =  Admin::where('user_id', Auth::user()->id)->first();
+        if(empty($admin)){
+         return response()->json(['message' => 'NO AUTHENTICATED'],201);
+        }
         $product->update($request->all());
         return response()->json(['message' => 'Product Updated Successfully','post' => $product],201);
 
@@ -64,6 +72,10 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        $admin =  Admin::where('user_id', Auth::user()->id)->first();
+        if(empty($admin)){
+         return response()->json(['message' => 'NO AUTHENTICATED'],201);
+        }
         $product->delete();
         return response()->json(['message' => 'Product DELETED Successfully','post' => $product],201);
     }
